Handle all possible errors in establishing connections, including address retrieving and resolving
